Move partial VFS invalidation test to own class
