+++ subprojects/snapshots/src/integTest/groovy/org/gradle/internal/vfs/VirtualFileSystemPartialInvalidationIntegrationTest.groovy
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2019 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.internal.vfs
+
+import org.gradle.integtests.fixtures.AbstractIntegrationSpec
+
+import static org.gradle.internal.service.scopes.VirtualFileSystemServices.VFS_PARTIAL_INVALIDATION_ENABLED_PROPERTY
+
+class VirtualFileSystemPartialInvalidationIntegrationTest extends AbstractIntegrationSpec {
+
+    def "incubating message is shown for partial invalidation"() {
+        buildFile << """
+            apply plugin: "java"
+        """
+        def incubatingMessage = "Partial virtual file system invalidation is an incubating feature"
+
+        when:
+        run("assemble", "-D${VFS_PARTIAL_INVALIDATION_ENABLED_PROPERTY}")
+        then:
+        outputContains(incubatingMessage)
+
+        when:
+        run("assemble")
+        then:
+        outputDoesNotContain(incubatingMessage)
+    }
+}
